move tasks (indexing, reindexing, mirror) to new dir
