+++ quetz/tasks/__init__.py
@@ -1,0 +1,1 @@
+from .common import execute_channel_action  # noqa
