+++ quetz/tasks/assertions.py
@@ -1,0 +1,6 @@
+def can_channel_synchronize(channel):
+    return channel.mirror_channel_url and (channel.mirror_mode == "mirror")
+
+
+def can_channel_reindex(channel):
+    return True
